requirements.txt
python-dotenv
fastapi
groq
pdfplumber
python-docx
python-multipart
uvicorn